Split warc_types into reader and writer files
